2007-10-04 [colin]	3.0.2cvs9
	* src/action.c
	* src/addrbook.c
	* src/addrindex.c
	* src/compose.c
	* src/exporthtml.c
	* src/exportldif.c
	* src/folder.c
	* src/main.c
	* src/matcher.c
	* src/mbox.c
	* src/messageview.c
	* src/mh.c
	* src/news.c
	* src/partial_download.c
	* src/pop.c
	* src/procmime.c
	* src/procmsg.c
	* src/toolbar.c
	* src/common/log.c
	* src/common/mgutils.c
	* src/common/plugin.c
	* src/common/utils.c
	* src/common/xml.c
	* src/common/xml.h
	* src/common/xmlprops.c
	* src/common/xmlprops.h
	* src/plugins/bogofilter/bogofilter.c
	* src/plugins/clamav/clamav_plugin.c
	* src/plugins/dillo_viewer/dillo_prefs.c
	* src/plugins/pgpcore/prefs_gpg.c
	* src/plugins/pgpinline/pgpinline.c
	* src/plugins/pgpmime/pgpmime.c
	* src/plugins/spamassassin/spamassassin.c
	* src/plugins/trayicon/trayicon_prefs.c
		Verify success on writes. Will fix possible strange
		failure on disk full.

+++ src/partial_download.c
@@ -214,8 +214,15 @@ static int partial_uidl_mark_mail(MsgInfo *msginfo, int download)
 			}
 		}
 		if (strcmp(tinfo->extradata->partial_recv, uidl)) {
-			fprintf(fpnew, "%s\t%ld\t%s\n", 
-				uidl, (long int) recv_time, partial_recv);
+			if (fprintf(fpnew, "%s\t%ld\t%s\n", 
+				uidl, (long int) recv_time, partial_recv) < 0) {
+				FILE_OP_ERROR(pathnew, "fprintf");
+				fclose(fpnew);
+				fclose(fp);
+				g_free(path);
+				g_free(pathnew);
+				goto bail;
+			}
 		} else {
 			gchar *stat = NULL;
 			if (download == POP3_PARTIAL_DLOAD_DLOAD) {
@@ -230,12 +237,25 @@ static int partial_uidl_mark_mail(MsgInfo *msginfo, int download)
 			else if (download == POP3_PARTIAL_DLOAD_DELE)
 				stat = g_strdup("0");
 			
-			fprintf(fpnew, "%s\t%ld\t%s\n", 
-				uidl, (long int) recv_time, stat);
+			if (fprintf(fpnew, "%s\t%ld\t%s\n", 
+				uidl, (long int) recv_time, stat) < 0) {
+				FILE_OP_ERROR(pathnew, "fprintf");
+				fclose(fpnew);
+				fclose(fp);
+				g_free(path);
+				g_free(pathnew);
+				goto bail;
+			}
 			g_free(stat);
 		}
 	}
-	fclose(fpnew);
+	if (fclose(fpnew) == EOF) {
+		FILE_OP_ERROR(pathnew, "fclose");
+		fclose(fp);
+		g_free(path);
+		g_free(pathnew);
+		goto bail;
+	}
 	fclose(fp);
 
 	move_file(pathnew, path, TRUE);
@@ -255,23 +275,47 @@ static int partial_uidl_mark_mail(MsgInfo *msginfo, int download)
 		goto bail;
 	}
 	
-	fprintf(fpnew, "SC-Marked-For-Download: %d\n", 
-			download);
+	if (fprintf(fpnew, "SC-Marked-For-Download: %d\n", 
+			download) < 0) {
+		FILE_OP_ERROR(pathnew, "fprintf");
+		fclose(fpnew);
+		fclose(fp);
+		g_free(pathnew);
+		goto bail;
+	}
 	while (fgets(buf, sizeof(buf)-1, fp) != NULL) {
 		if(strlen(buf) > strlen("SC-Marked-For-Download: x\n")
 		&& !strncmp(buf, "SC-Marked-For-Download:", 
 		            strlen("SC-Marked-For-Download:"))) {
-			fprintf(fpnew, "%s", 
-			 buf+strlen("SC-Marked-For-Download: x\n"));
+			if (fprintf(fpnew, "%s", 
+			 buf+strlen("SC-Marked-For-Download: x\n")) < 0) {
+				FILE_OP_ERROR(pathnew, "fprintf");
+				fclose(fpnew);
+				fclose(fp);
+				g_free(pathnew);
+				goto bail;
+			}
 			continue;
 		} else if (strlen(buf) == strlen("SC-Marked-For-Download: x\n")
 		&& !strncmp(buf, "SC-Marked-For-Download:", 
 		            strlen("SC-Marked-For-Download:"))) {
 			continue;
 		}
-		fprintf(fpnew, "%s", buf);
+		if (fprintf(fpnew, "%s", buf) < 0) {
+			FILE_OP_ERROR(pathnew, "fprintf");
+			fclose(fpnew);
+			fclose(fp);
+			g_free(pathnew);
+			goto bail;
+		}
 	}
-	fclose(fpnew);
+	if (fclose(fpnew) == EOF) {
+		FILE_OP_ERROR(pathnew, "fclose");
+		fclose(fp);
+		g_free(pathnew);
+		goto bail;
+	}
+
 	fclose(fp);
 	g_unlink(filename);
 	rename(pathnew, filename);
